requirements.txt
python-dotenv
numpy
pandas
pymongo
certifi
pymongo[srv]==3.11

# -e .